# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "\u003ch3\u003eThis server is a bidirectional queue using gRCP. Manager section\u003c/h3\u003e",
        "title": "dominus-project server",
        "contact": {
            "name": "MBI",
            "url": "https://www.pr0c0d3.com/",
            "email": "ingmbi8807@gmail.com"
        },
        "version": "1.1.0"
    },
    "paths": {
        "/health": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "\u003ch3\u003eget healthCeck\u003c/h3\u003e",
                "tags": [
                    "Monitor"
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Response body {message: error}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "\u003ch3\u003egets metrics\u003c/h3\u003e",
                "tags": [
                    "Monitor"
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Response body {message: error}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/logs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "\u003ch3\u003egets all logs fron the database using page and size (required)\u003c/h3\u003e",
                "tags": [
                    "Logs"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Response body {message: error}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/logs-backup": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "\u003ch3\u003eGets all selected items from the database for making a backup\u003c/h3\u003e",
                "tags": [
                    "Logs"
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Error response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/subscribers/{name}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "\u003ch3\u003eupdate  subscribers\u003c/h3\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "topic name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Subscribers",
                        "name": "subscribers",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "406": {
                        "description": "Response body {message: error}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/topics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "\u003ch3\u003eget topics info\u003c/h3\u003e",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "responses": {
                    "200": {
                        "description": "Topic response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "406": {
                        "description": "Response body {message: error}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "\u003ch3\u003eadd a new topic\u003c/h3\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "parameters": [
                    {
                        "description": "Topic Info",
                        "name": "topic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerTopic"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "406": {
                        "description": "Response body {message: error}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/topics/{name}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "\u003ch3\u003edelete a topic\u003c/h3\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "topic name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "406": {
                        "description": "Response body {message: error}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.SwaggerTopic": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Name of the topic, e.g., prod.topic, monitoring, etc.",
                    "type": "string",
                    "example": "prod.topic"
                },
                "subscribers": {
                    "description": "List of subscriber URLs",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[localhost:8080",
                        " localhost:8081]"
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    }
}